Update Assignment.java
it will not worked.. so that i make some changes into it. Its gave an error in "ValuesOutOfRangeException" and "NegativeValuesException"  these classes.

I have to make my own classes to handle the negetive integer values. for that reason i create a new class "NegativeValuesException" which extends from Exception class in outside of main class so that every class can get  accessed.
I have to make my own classes to handle the out of range values. for that reason i create a new class "ValuesOutOfRangeException" which extends from Exception class in outside of main class so that every class can get accessed.

Thank you 😃